added heatmaps to the cli
