+++ cea/interfaces/arcgis/CityEnergyAnalyst.py
@@ -12,6 +12,10 @@ import os
 import subprocess
 import tempfile
 
+import cea
+import cea.inputlocator
+import cea.plots
+
 __author__ = "Daren Thomas"
 __copyright__ = "Copyright 2016, Architecture and Building Systems - ETH Zurich"
 __credits__ = ["Daren Thomas"]
@@ -27,9 +31,8 @@ class Toolbox(object):
     def __init__(self):
         self.label = 'City Energy Analyst'
         self.alias = 'cea'
-        # self.tools = [HeatmapsTool, RadiationTool]
         self.tools = [DemandTool, DataHelperTool, BenchmarkGraphsTool, EmissionsTool, EmbodiedEnergyTool, MobilityTool,
-                      DemandGraphsTool, ScenarioPlotsTool, RadiationTool]
+                      DemandGraphsTool, ScenarioPlotsTool, RadiationTool, HeatmapsTool]
 
 
 class DemandTool(object):
@@ -613,3 +616,79 @@ def run_cli(scenario_path=None, *args):
     stdout, stderr = process.communicate()
     add_message(stdout)
     add_message(stderr)
+
+
+class HeatmapsTool(object):
+    def __init__(self):
+        self.label = 'Heatmaps'
+        self.description = 'Create heatmap data layers'
+        self.canRunInBackground = False
+
+    def getParameterInfo(self):
+        import arcpy
+        scenario_path = arcpy.Parameter(
+            displayName="Path to the scenario",
+            name="scenario_path",
+            datatype="DEFolder",
+            parameterType="Required",
+            direction="Input")
+        path_variables = arcpy.Parameter(
+            displayName="Choose the file to analyse",
+            name="path_variables",
+            datatype="String",
+            parameterType="Required",
+            direction="Input")
+        path_variables.filter.list = []
+        analysis_fields = arcpy.Parameter(
+            displayName="Variables to analyse",
+            name="analysis_fields",
+            datatype="String",
+            parameterType="Required",
+            multiValue=True,
+            direction="Input")
+        analysis_fields.filter.list = []
+        analysis_fields.parameterDependencies = ['path_variables']
+
+        return [scenario_path, path_variables, analysis_fields]
+
+
+    def updateParameters(self, parameters):
+        # scenario_path
+        scenario_path = parameters[0].valueAsText
+        if not os.path.exists(scenario_path):
+            parameters[0].setErrorMessage('Scenario folder not found: %s' % scenario_path)
+            return
+        # path_variables
+        file_names = [os.path.basename(_cli_output(scenario_path, 'locate', 'get_total_demand'))]
+        file_names.extend(
+            [f for f in os.listdir(_cli_output(scenario_path, 'locate', 'get_lca_emissions_results_folder'))
+             if f.endswith('.csv')])
+        path_variables = parameters[1]
+        if not path_variables.value or path_variables.value not in file_names:
+            path_variables.filter.list = file_names
+            path_variables.value = file_names[0]
+        # analysis_fields
+        analysis_fields = parameters[2]
+        if path_variables.value == file_names[0]:
+            file_to_analyze = _cli_output(scenario_path, 'locate', 'get_total_demand')
+        else:
+            file_to_analyze = os.path.join(_cli_output(scenario_path, 'locate', 'get_lca_emissions_results_folder'),
+                                           path_variables.value)
+        import pandas as pd
+        df = pd.read_csv(file_to_analyze)
+        fields = df.columns.tolist()
+        fields.remove('Name')
+        analysis_fields.filter.list = list(fields)
+        return
+
+    def execute(self, parameters, _):
+        scenario_path = parameters[0].valueAsText
+        file_to_analyze = parameters[1].valueAsText
+        analysis_fields = parameters[2].valueAsText.split(';')
+
+        if file_to_analyze == os.path.basename(_cli_output(scenario_path, 'locate', 'get_total_demand')):
+            file_to_analyze = _cli_output(scenario_path, 'locate', 'get_total_demand')
+        else:
+            file_to_analyze = os.path.join(_cli_output(scenario_path, 'locate', 'get_lca_emissions_results_folder'),
+                                           file_to_analyze)
+        run_cli(scenario_path, 'heatmaps', '--file-to-analyze', file_to_analyze, '--analysis-fields', *analysis_fields)
